# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Solar Panel ERP API",
...
    "version": "1.0.0"
  },
  "paths": {
...
    "/inventory/movements": {
      "post": {
        "tags": [
          "Inventory ledger"
        ],
        "summary": "Post double-entry inventory movement",
        "description": "Posts an append-only double-entry physical inventory movement with one or more lines.\n\n- Movement types: RECEIPT, SHIPMENT, TRANSFER, ADJUSTMENT_IN, ADJUSTMENT_OUT, PRODUCTION_CONSUMPTION, PRODUCTION_OUTPUT.\n- Database triggers enforce that source location balances never drop below zero.\n- Safe retries via Idempotency-Key header.",
        "operationId": "post_movement_inventory_movements_post",
        "parameters": [
          {
            "name": "idempotency-key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "minLength": 1,
                  "maxLength": 100,
                  "pattern": "^\\S+$"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Idempotency-Key"
            }
          },
          {
            "name": "x-organization-id",
            "in": "header",
            "required": true,
            "schema": {
              "type": "integer",
              "maximum": 9223372036854775807,
              "exclusiveMinimum": 0,
              "title": "X-Organization-Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MovementCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MovementRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/inventory/items/{revision_id}/on-hand": {
      "get": {
        "tags": [
          "Inventory ledger"
        ],
        "summary": "Get physical on-hand stock",
        "description": "Calculates real-time physical on-hand inventory across all locations for an item revision.\n\n- Formula: Incoming - Outgoing across append-only movement lines.\n- Includes warehouse stock, production bins, and quarantine/damaged stock.",
        "operationId": "get_on_hand_inventory_items__revision_id__on_hand_get",
        "parameters": [
          {
            "name": "revision_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "maximum": 9223372036854775807,
              "exclusiveMinimum": 0,
              "title": "Revision Id"
            }
          },
          {
            "name": "x-organization-id",
            "in": "header",
            "required": true,
            "schema": {
              "type": "integer",
              "maximum": 9223372036854775807,
              "exclusiveMinimum": 0,
              "title": "X-Organization-Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OnHandRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/inventory/items/{revision_id}/availability": {
      "get": {
        "tags": [
          "Inventory ledger"
        ],
        "summary": "Get available stock",
        "description": "Calculates available-to-promise inventory accounting for active reservations.\n\n- Formula: On Hand - Active Reservations.\n- All three totals and the location breakdown cover the same eligible locations.\n- Reservable Locations Invariant: Only active locations of type WAREHOUSE, BIN, and PRODUCTION count.\n- Quarantine (QUARANTINE), damaged, and transit locations are strictly excluded.\n- Use /on-hand for physical stock across all locations, including excluded stock.",
        "operationId": "get_availability_inventory_items__revision_id__availability_get",
        "parameters": [
          {
            "name": "revision_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "maximum": 9223372036854775807,
              "exclusiveMinimum": 0,
              "title": "Revision Id"
            }
          },
          {
            "name": "x-organization-id",
            "in": "header",
            "required": true,
            "schema": {
              "type": "integer",
              "maximum": 9223372036854775807,
              "exclusiveMinimum": 0,
              "title": "X-Organization-Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AvailabilityRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/inventory/reservations": {
      "post": {
        "tags": [
          "Inventory ledger"
        ],
        "summary": "Create stock reservation",
        "description": "Creates an active stock reservation against available inventory.\n\n- Option A Traceability: Must reference either a sales_order_line_id or production_order_material_id.\n- Rejects if quantity exceeds currently available stock at the location.",
        "operationId": "post_reservation_inventory_reservations_post",
        "parameters": [
          {
            "name": "x-organization-id",
            "in": "header",
            "required": true,
            "schema": {
              "type": "integer",
              "maximum": 9223372036854775807,
              "exclusiveMinimum": 0,
              "title": "X-Organization-Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReservationCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReservationRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/inventory/reservations/{reservation_id}": {
      "get": {
        "tags": [
          "Inventory ledger"
        ],
        "summary": "Get reservation details",
        "description": "Retrieves reservation details, current status (ACTIVE, RELEASED, or CONSUMED), and ownership.",
        "operationId": "get_reservation_inventory_reservations__reservation_id__get",
        "parameters": [
          {
            "name": "reservation_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "maximum": 9223372036854775807,
              "exclusiveMinimum": 0,
              "title": "Reservation Id"
            }
          },
          {
            "name": "x-organization-id",
            "in": "header",
            "required": true,
            "schema": {
              "type": "integer",
              "maximum": 9223372036854775807,
              "exclusiveMinimum": 0,
              "title": "X-Organization-Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReservationRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Inventory ledger"
        ],
        "summary": "Release reservation",
        "description": "Explicitly releases an active stock reservation (status -> RELEASED).\n\n- Restores available stock to the location.\n- Row is preserved in database for audit history; never physically deleted.",
        "operationId": "delete_reservation_inventory_reservations__reservation_id__delete",
        "parameters": [
          {
            "name": "reservation_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "maximum": 9223372036854775807,
              "exclusiveMinimum": 0,
              "title": "Reservation Id"
            }
          },
          {
            "name": "x-organization-id",
            "in": "header",
            "required": true,
            "schema": {
              "type": "integer",
              "maximum": 9223372036854775807,
              "exclusiveMinimum": 0,
              "title": "X-Organization-Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReservationRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/inventory/reservations/{reservation_id}/release": {
      "post": {
        "tags": [
          "Inventory ledger"
        ],
        "summary": "Release reservation",
        "description": "Explicitly releases an active stock reservation (status -> RELEASED).\n\n- Restores available stock to the location.\n- Row is preserved in database for audit history; never physically deleted.",
        "operationId": "delete_reservation_inventory_reservations__reservation_id__release_post",
        "parameters": [
          {
            "name": "reservation_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "maximum": 9223372036854775807,
              "exclusiveMinimum": 0,
              "title": "Reservation Id"
            }
          },
          {
            "name": "x-organization-id",
            "in": "header",
            "required": true,
            "schema": {
              "type": "integer",
              "maximum": 9223372036854775807,
              "exclusiveMinimum": 0,
              "title": "X-Organization-Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReservationRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/inventory/reservations/{reservation_id}/consume": {
      "post": {
        "tags": [
          "Inventory ledger"
        ],
        "summary": "Consume reservation",
        "description": "Atomically consumes an active reservation and posts its physical stock issue.\n\n- Sales allocations post SHIPMENT; material allocations post PRODUCTION_CONSUMPTION.\n- Retrying a consumed reservation returns its existing state without another issue.\n- Production order completion owns its own atomic material and output workflow.\n- Terminal state; consumed reservations cannot be released.",
        "operationId": "consume_reservation_inventory_reservations__reservation_id__consume_post",
        "parameters": [
          {
            "name": "reservation_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "maximum": 9223372036854775807,
              "exclusiveMinimum": 0,
              "title": "Reservation Id"
            }
          },
          {
            "name": "x-organization-id",
            "in": "header",
            "required": true,
            "schema": {
              "type": "integer",
              "maximum": 9223372036854775807,
              "exclusiveMinimum": 0,
              "title": "X-Organization-Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReservationRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AvailabilityRead": {
        "properties": {
          "organization_id": {
            "type": "integer",
            "title": "Organization Id"
          },
          "item_revision_id": {
            "type": "integer",
            "title": "Item Revision Id"
          },
          "item_id": {
            "type": "integer",
            "title": "Item Id"
          },
          "sku": {
            "type": "string",
            "title": "Sku"
          },
          "revision_code": {
            "type": "string",
            "title": "Revision Code"
          },
          "base_uom": {
            "type": "string",
            "title": "Base Uom"
          },
          "on_hand": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "On Hand"
          },
          "reserved": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Reserved"
          },
          "available": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Available"
          },
          "locations": {
            "items": {
              "$ref": "#/components/schemas/LocationAvailability"
            },
            "type": "array",
            "title": "Locations"
          }
        },
        "type": "object",
        "required": [
          "organization_id",
          "item_revision_id",
          "item_id",
          "sku",
          "revision_code",
          "base_uom",
          "on_hand",
          "reserved",
          "available",
          "locations"
        ],
        "title": "AvailabilityRead"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "LocationBalance": {
        "properties": {
          "location_id": {
            "type": "integer",
            "title": "Location Id"
          },
          "code": {
            "type": "string",
            "title": "Code"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "incoming_quantity": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Incoming Quantity"
          },
          "outgoing_quantity": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Outgoing Quantity"
          },
          "on_hand_quantity": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "On Hand Quantity"
          }
        },
        "type": "object",
        "required": [
          "location_id",
          "code",
          "name",
          "incoming_quantity",
          "outgoing_quantity",
          "on_hand_quantity"
        ],
        "title": "LocationBalance"
      },
...
      "MovementCreate": {
        "properties": {
          "movement_type": {
            "type": "string",
            "enum": [
              "RECEIPT",
              "TRANSFER",
              "ADJUSTMENT_IN",
              "ADJUSTMENT_OUT"
            ],
            "title": "Movement Type"
          },
          "reference": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 100
              },
              {
                "type": "null"
              }
            ],
            "title": "Reference"
          },
          "occurred_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Occurred At"
          },
          "lines": {
            "items": {
              "$ref": "#/components/schemas/MovementLineCreate"
            },
            "type": "array",
            "minItems": 1,
            "title": "Lines"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "movement_type",
          "lines"
        ],
        "title": "MovementCreate"
      },
      "MovementLineCreate": {
        "properties": {
          "item_revision_id": {
            "type": "integer",
            "maximum": 9.223372036854776e+18,
            "exclusiveMinimum": 0.0,
            "title": "Item Revision Id"
          },
          "from_location_id": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 9.223372036854776e+18,
                "exclusiveMinimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "From Location Id"
          },
          "to_location_id": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 9.223372036854776e+18,
                "exclusiveMinimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "To Location Id"
          },
          "quantity": {
            "anyOf": [
              {
                "type": "number",
                "exclusiveMinimum": 0.0
              },
              {
                "type": "string",
                "pattern": "^(?!^[-+.]*$)[+-]?0*(?:\\d{0,12}|(?=[\\d.]{1,19}0*$)\\d{0,12}\\.\\d{0,6}0*$)"
              }
            ],
            "title": "Quantity"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "item_revision_id",
          "quantity"
        ],
        "title": "MovementLineCreate"
      },
      "MovementLineRead": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "organization_id": {
            "type": "integer",
            "title": "Organization Id"
          },
          "movement_id": {
            "type": "integer",
            "title": "Movement Id"
          },
          "item_revision_id": {
            "type": "integer",
            "title": "Item Revision Id"
          },
          "from_location_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "From Location Id"
          },
          "to_location_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "To Location Id"
          },
          "quantity": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Quantity"
          }
        },
        "type": "object",
        "required": [
          "id",
          "organization_id",
          "movement_id",
          "item_revision_id",
          "from_location_id",
          "to_location_id",
          "quantity"
        ],
        "title": "MovementLineRead"
      },
      "MovementRead": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "organization_id": {
            "type": "integer",
            "title": "Organization Id"
          },
          "movement_type": {
            "type": "string",
            "enum": [
              "RECEIPT",
              "TRANSFER",
              "ADJUSTMENT_IN",
              "ADJUSTMENT_OUT"
            ],
            "title": "Movement Type"
          },
          "reference": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reference"
          },
          "occurred_at": {
            "type": "string",
            "format": "date-time",
            "title": "Occurred At"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "lines": {
            "items": {
              "$ref": "#/components/schemas/MovementLineRead"
            },
            "type": "array",
            "title": "Lines"
          }
        },
        "type": "object",
        "required": [
          "id",
          "organization_id",
          "movement_type",
          "reference",
          "occurred_at",
          "created_at",
          "lines"
        ],
        "title": "MovementRead"
      },
      "OnHandRead": {
        "properties": {
          "organization_id": {
            "type": "integer",
            "title": "Organization Id"
          },
          "item_revision_id": {
            "type": "integer",
            "title": "Item Revision Id"
          },
          "item_id": {
            "type": "integer",
            "title": "Item Id"
          },
          "sku": {
            "type": "string",
            "title": "Sku"
          },
          "revision_code": {
            "type": "string",
            "title": "Revision Code"
          },
          "base_uom": {
            "type": "string",
            "title": "Base Uom"
          },
          "incoming_quantity": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Incoming Quantity"
          },
          "outgoing_quantity": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Outgoing Quantity"
          },
          "on_hand_quantity": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "On Hand Quantity"
          },
          "locations": {
            "items": {
              "$ref": "#/components/schemas/LocationBalance"
            },
            "type": "array",
            "title": "Locations"
          }
        },
        "type": "object",
        "required": [
          "organization_id",
          "item_revision_id",
          "item_id",
          "sku",
          "revision_code",
          "base_uom",
          "incoming_quantity",
          "outgoing_quantity",
          "on_hand_quantity",
          "locations"
        ],
        "title": "OnHandRead"
      },
...
      "ReservationCreate": {
        "properties": {
          "item_revision_id": {
            "type": "integer",
            "maximum": 9.223372036854776e+18,
            "exclusiveMinimum": 0.0,
            "title": "Item Revision Id"
          },
          "location_id": {
            "type": "integer",
            "maximum": 9.223372036854776e+18,
            "exclusiveMinimum": 0.0,
            "title": "Location Id"
          },
          "quantity": {
            "anyOf": [
              {
                "type": "number",
                "exclusiveMinimum": 0.0
              },
              {
                "type": "string",
                "pattern": "^(?!^[-+.]*$)[+-]?0*(?:\\d{0,12}|(?=[\\d.]{1,19}0*$)\\d{0,12}\\.\\d{0,6}0*$)"
              }
            ],
            "title": "Quantity"
          },
          "sales_order_line_id": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 9.223372036854776e+18,
                "exclusiveMinimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Sales Order Line Id"
          },
          "production_order_material_id": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 9.223372036854776e+18,
                "exclusiveMinimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Production Order Material Id"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "item_revision_id",
          "location_id",
          "quantity"
        ],
        "title": "ReservationCreate"
      },
      "ReservationRead": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "organization_id": {
            "type": "integer",
            "title": "Organization Id"
          },
          "item_revision_id": {
            "type": "integer",
            "title": "Item Revision Id"
          },
          "location_id": {
            "type": "integer",
            "title": "Location Id"
          },
          "sales_order_line_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sales Order Line Id"
          },
          "production_order_material_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Production Order Material Id"
          },
          "quantity": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Quantity"
          },
          "status": {
            "type": "string",
            "enum": [
              "ACTIVE",
              "CONSUMED",
              "RELEASED"
            ],
            "title": "Status"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "organization_id",
          "item_revision_id",
          "location_id",
          "sales_order_line_id",
          "production_order_material_id",
          "quantity",
          "status",
          "created_at",
          "updated_at"
        ],
        "title": "ReservationRead"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  },
...
}